geoq/requirements.txt
Django==1.5.4
Django-Select2==4.2.1
#PIL==1.1.7
Paver==1.2.1
South==0.8.2
django-appconf==0.6
#django-badges==0.1.6
django-bootstrap-toolkit==2.15.0
django-compressor==1.3
django-geoexplorer==3.0.4
django-guardian==1.1.1
django-leaflet==0.8.0
django-reversion==1.7.1
django-singleton==0.1.7
django-userena==1.2.4
easy-thumbnails==1.4
jsonfield==0.9.17
psycopg2==2.5.1
requests==1.2.3
six==1.4.1